Rename Repository to Singleton
Rename Repository to Singleton
