+++ app/src/main/java/com/cornellappdev/coffee_chats_android/fragments/EditInterestsGroupsFragment.kt
@@ -13,7 +13,7 @@ import com.cornellappdev.coffee_chats_android.adapters.UserFieldAdapter
 import com.cornellappdev.coffee_chats_android.adapters.UserFieldAdapter.ItemColor
 import com.cornellappdev.coffee_chats_android.models.UserField
 import com.cornellappdev.coffee_chats_android.models.UserField.Category
-import com.cornellappdev.coffee_chats_android.repositories.UserRepository
+import com.cornellappdev.coffee_chats_android.singletons.UserSingleton
 import kotlinx.android.synthetic.main.fragment_edit_interests.*
 import kotlinx.android.synthetic.main.fragment_edit_interests.view.*
 import kotlinx.android.synthetic.main.interest_group_list_with_header.view.*
@@ -63,7 +63,7 @@ class EditInterestsGroupsFragment : Fragment(), OnFilledOutObservable {
         super.onResume()
         selectedItems.clear()
         if (isInterest) {
-            val userInterests = UserRepository.user.interests
+            val userInterests = UserSingleton.user.interests
             for (interest in userInterests) {
                 val item = UserField(
                     interest.name,
@@ -74,7 +74,7 @@ class EditInterestsGroupsFragment : Fragment(), OnFilledOutObservable {
                 selectedItems.add(item)
             }
         } else {
-            val userGroups = UserRepository.user.groups
+            val userGroups = UserSingleton.user.groups
             for (group in userGroups) {
                 val item = UserField(group.name, drawableUrl = group.imageUrl, id = group.id)
                 selectedItems.add(item)
@@ -140,9 +140,9 @@ class EditInterestsGroupsFragment : Fragment(), OnFilledOutObservable {
     private fun removeItem(pos: Int) {
         val removedItemId = selectedItems.removeAt(pos).id
         if (isInterest) {
-            UserRepository.removeInterest(removedItemId)
+            UserSingleton.removeInterest(removedItemId)
         } else {
-            UserRepository.removeGroup(removedItemId)
+            UserSingleton.removeGroup(removedItemId)
         }
         selectedItemsAdapter.notifyDataSetChanged()
         toggleSaveButton()
